backend/requirements.txt
Django>=4.0,<5.0  # Укажите версию Django, которую вы используете
djangorestframework>=3.14.0  # Укажите версию Django REST Framework
# Добавьте дополнительные зависимости по мере необходимости
# Например, если вы используете CORS:
django-cors-headers>=3.14.0
# Если вы планируете использовать базу данных PostgreSQL:
psycopg2>=2.9.1
# Если вам нужно работать с изображениями:
Pillow>=9.0.0
# Для тестирования можно добавить:
pytest>=7.0.0
pytest-django>=4.5.0